feat(AC): moved range date handler to own file
